TINY-8152: Clean up and move SelectionTargets back to table plugin
