+++ modules/tinymce/src/core/main/ts/keyboard/TabKey.ts
@@ -13,7 +13,7 @@ import * as MatchKeys from './MatchKeys';
 import * as TableNavigation from './TableNavigation';
 
 const tableTabNavigation = (editor: Editor): MatchKeys.KeyPattern[] => {
-  if (allowTableTabNavigation) {
+  if (allowTableTabNavigation(editor)) {
     return [
       { keyCode: VK.TAB, action: MatchKeys.action(TableNavigation.handleTab, editor, true) },
       { keyCode: VK.TAB, shiftKey: true, action: MatchKeys.action(TableNavigation.handleTab, editor, false) },
