+++ modules/tinymce/src/models/dom/main/ts/table/main/ts/Table.ts
@@ -5,8 +5,6 @@
  * For commercial licenses see https://www.tiny.cloud/
  */
 
-// import { Selections } from '@ephox/darwin';
-// import { Arr, Fun } from '@ephox/katamari';
 import { SugarElement } from '@ephox/sugar';
 
 import Editor from 'tinymce/core/api/Editor';
@@ -17,39 +15,19 @@ import { TableActions } from './actions/TableActions';
 import { Api, getApi } from './api/Api';
 import * as Commands from './api/Commands';
 import * as QueryCommands from './api/QueryCommands';
-// import { hasTabNavigation } from './api/Settings';
 import { Clipboard as FakeClipboard } from './core/Clipboard';
 import * as TableFormats from './core/TableFormats';
-// import * as Util from './core/Util';
-// import * as TabContext from './queries/TabContext';
-// import CellSelection from './selection/CellSelection';
 import { ephemera } from './selection/Ephemera';
-import { getSelectionTargets } from './selection/SelectionTargets';
-// import { getSelectionCell } from './selection/TableSelection';
+// import { getSelectionTargets } from './selection/SelectionTargets';
 
 export interface PatchedSelections {
   readonly get: () => SugarElement<HTMLTableCellElement>[];
 }
 
-// const patchSelections = (selections: Selections): PatchedSelections => {
-//   return {
-//     get: () => selections.get().fold(Fun.constant([]), Fun.identity, Arr.pure)
-//   };
-// };
-
 const setupTable = (editor: Editor): Api => {
-  // Move selection and resizing logic to actual core
-  // const oldSelections = Selections(() => Util.getBody(editor), () => getSelectionCell(Util.getSelectionStart(editor), Util.getIsRoot(editor)), ephemera.selectedSelector);
-  // const selections = patchSelections(oldSelections);
-
-  const selectionTargets = getSelectionTargets(editor);
+  // const selectionTargets = getSelectionTargets(editor);
+  // TODO: Try and move into core and resize bars out of snooker so that we don't need to pass lazyWire into Snooker. Maybe be able to use an event system instead?
   const resizeHandler = getResizeHandler(editor);
-  // const resizeHandler = editor.selection.tableResizeHandler;
-  // console.log(resizeHandler);
-  // TODO: I don't think we want CellSelection here as the selection should be in core but leave here for now
-  // const cellSelection = CellSelection(editor, selectionTargets, resizeHandler.lazyResize);
-  // const cellSelection = CellSelection(editor, resizeHandler.lazyResize);
-  // const actions = TableActions(editor, cellSelection, resizeHandler.lazyWire);
 
   // TODO: To solve resizeHandler issue, could put register all of this on init to allow Editor to initialise ResizeHandler but doesn't seem like a great solution
 
@@ -68,22 +46,11 @@ const setupTable = (editor: Editor): Api => {
     TableFormats.registerFormats(editor);
   });
 
-  // TODO: Move to core - keyboard overrides
-  // if (hasTabNavigation(editor)) {
-  // editor.on('keydown', (e: KeyboardEvent) => {
-  // TabContext.handle(e, editor, cellSelection);
-  // TabContext.handle(e, editor);
-  // });
-  // }
-
   editor.on('remove', () => {
     resizeHandler.destroy();
   });
 
-  // TODO: Attempt making the API just in the internal APIs
-  // Maybe add ephemera to the API as well
-  // return getApi(clipboard, resizeHandler, selectionTargets, cellSelection);
-  return getApi(clipboard, resizeHandler, selectionTargets);
+  return getApi(clipboard, resizeHandler);
 };
 
 export {
