+++ modules/tinymce/src/models/dom/main/ts/table/main/ts/api/Api.ts
@@ -10,21 +10,21 @@
 import { ResizeHandler } from '../actions/ResizeHandler';
 import { Clipboard as FakeClipboard } from '../core/Clipboard';
 // import { CellSelectionApi } from '../selection/CellSelection';
-import { SelectionTargets } from '../selection/SelectionTargets';
+// import { SelectionTargets } from '../selection/SelectionTargets';
 
 export interface Api {
   readonly fakeClipboard: FakeClipboard;
   readonly resizeHandler: ResizeHandler;
-  readonly selectionTargets: SelectionTargets;
+  // readonly selectionTargets: SelectionTargets;
   // readonly selections: Selections;
   // readonly cellSelection: CellSelectionApi;
 }
 
 // const getApi = (clipboard: FakeClipboard, resizeHandler: ResizeHandler, selectionTargets: SelectionTargets, cellSelection: CellSelectionApi): Api => ({
-const getApi = (clipboard: FakeClipboard, resizeHandler: ResizeHandler, selectionTargets: SelectionTargets): Api => ({
+const getApi = (clipboard: FakeClipboard, resizeHandler: ResizeHandler): Api => ({
   fakeClipboard: clipboard,
   resizeHandler,
-  selectionTargets,
+  // selectionTargets,
   // selections,
   // cellSelection
 });
