+++ modules/tinymce/src/plugins/table/main/ts/api/Api.ts
@@ -10,7 +10,6 @@ import { SugarElement, SugarElements } from '@ephox/sugar';
 
 import Editor from 'tinymce/core/api/Editor';
 import { ResizeHandler } from 'tinymce/models/dom/table/main/ts/actions/ResizeHandler';
-// import { Api as TableModelApi } from 'tinymce/models/dom/table/main/ts/api/Api';
 import { Clipboard as FakeClipboard } from 'tinymce/models/dom/table/main/ts/core/Clipboard';
 
 import { insertTableWithDataValidation } from '../actions/InsertTable';
@@ -49,7 +48,6 @@ const getApi = (editor: Editor, clipboard: FakeClipboard, selectionTargets: Sele
   getClipboardRows: getClipboardElements(clipboard.getRows),
   setClipboardCols: setClipboardElements(clipboard.setColumns),
   getClipboardCols: getClipboardElements(clipboard.getColumns),
-  // resizeHandler: editor.selection.tableResizeHandler,
   resizeHandler: editor.model.table.resizeHandler,
   selectionTargets,
 });
